requirements.txt
Pillow
tinify
tqdm